HeightTile und Heightmap, Einfüge - und Suchmethoden -ungetestet-

+++ java/src/kit/route/a/lot/heightinfo/IHeightmap.java
@@ -25,7 +25,7 @@ public interface IHeightmap {
      *            -
      * @return float
      */
-    public float getHeight(Coordinates pos);
+    public int getHeight(Coordinates pos);
 
     /**
      * Operation addHeightTile
